+++ cartridge/webui/api-vshard.lua
@@ -20,7 +20,28 @@ local gql_type_vsgroup = gql_types.object({
             kind = gql_types.boolean.nonNull,
             description = 'Whethe the group is ready to operate',
         },
-
+        rebalancer_max_receiving = {
+            kind = gql_types.int.nonNull,
+            description =
+                'The maximum number of buckets that can be received in parallel by a single replica set ' ..
+                'in the storage group'
+        },
+        collect_lua_garbage = {
+            kind = gql_types.boolean.nonNull,
+            description = 'If set to true, the Lua collectgarbage() function is called periodically'
+        },
+        sync_timeout = {
+            kind = gql_types.float.nonNull,
+            description = 'Timeout to wait for synchronization of the old master with replicas before demotion'
+        },
+        collect_bucket_garbage_interval = {
+            kind = gql_types.float.nonNull,
+            description = 'The interval between garbage collector actions, in seconds'
+        },
+        rebalancer_disbalance_threshold = {
+            kind = gql_types.float.nonNull,
+            description = 'A maximum bucket disbalance threshold, in percent'
+        },
     }
 })
 
@@ -65,6 +86,19 @@ local function get_vshard_groups()
     return ret
 end
 
+local function edit_vshard_options(_, args)
+    local group_name = args.name
+    args.name = nil
+
+    local _, err = vshard_utils.edit_vshard_options(group_name, args)
+    if err ~= nil then
+        return nil, err
+    end
+
+    local group = vshard_utils.get_known_groups()[group_name]
+    group.name = group_name
+    return group
+end
 
 local function init(graphql)
     graphql.add_mutation({
@@ -110,6 +144,21 @@ local function init(graphql)
         kind = gql_types.list(gql_type_vsgroup.nonNull).nonNull,
         callback = module_name .. '.get_vshard_groups',
     })
+
+    graphql.add_mutation({
+        prefix = 'cluster',
+        name = 'edit_vshard_options',
+        args = {
+            name = gql_types.string.nonNull,
+            rebalancer_max_receiving = gql_types.int,
+            collect_lua_garbage = gql_types.boolean,
+            sync_timeout = gql_types.float,
+            collect_bucket_garbage_interval = gql_types.float,
+            rebalancer_disbalance_threshold = gql_types.float,
+        },
+        kind = gql_type_vsgroup.nonNull,
+        callback = module_name .. '.edit_vshard_options',
+    })
 end
 
 return {
@@ -118,4 +167,5 @@ return {
     get_vshard_bucket_count = get_vshard_bucket_count,
     get_vshard_known_groups = get_vshard_known_groups,
     get_vshard_groups = get_vshard_groups,
+    edit_vshard_options = edit_vshard_options
 }